requirements.txt
alembic==1.8.1
blinker==1.5
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
csrf==0.1b1
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
protobuf==4.21.2
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
python-dotenv==0.20.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
urllib3==1.26.11
uuid==1.30
Werkzeug==2.1.2
WTForms==3.0.1
